Force an Object to Belong to a Class

Coerce an object to a given class.

Value

Object of new class.

Description

as(object) returns the version of this object coerced to be the given Class. When used in the replacement form on the left of an assignment, the portion of the object corresponding to Class is replaced by value.

The operation of as() in either form depends on the definition of coerce methods. Methods are defined automatically when the two classes are related by inheritance; that is, when one of the classes is a subclass of the other.

Coerce methods are also predefined for basic classes (including all the types of vectors, functions and a few others).

Beyond these two sources of methods, further methods are defined by calls to the setAs function. See that documentation also for details of how coerce methods work. Use showMethods(coerce) for a list of all currently defined methods, as in the example below.

Basic Coercion Methods

Methods are pre-defined for coercing any object to one of the basic datatypes. For example, as(x, "numeric") uses the existing as.numeric function. These and all other existing methods can be listed as shown in the example.

References

Chambers, John M. (2016) Extending R, Chapman & Hall. (Chapters 9 and 10.)

See also

Examples

## Show all the existing methods for as()
showMethods("coerce")

#> Function: coerce (package methods)
#> from="AAString", to="MaskedAAString"
#> from="abIndex", to="integer"
#> from="abIndex", to="numeric"
#> from="abIndex", to="seqMat"
#> from="abIndex", to="vector"
#> from="AnnotatedDataFrame", to="data.frame"
#> from="ANY", to="AAStringSet"
#> from="ANY", to="array"
#> from="ANY", to="AsIs"
#> from="ANY", to="BStringSet"
#> from="ANY", to="call"
#> from="ANY", to="character"
#> from="ANY", to="CharacterList"
#> from="ANY", to="complex"
#> from="ANY", to="ComplexList"
#> from="ANY", to="CompressedCharacterList"
#> from="ANY", to="CompressedComplexList"
#> from="ANY", to="CompressedDataFrameList"
#> from="ANY", to="CompressedFactorList"
#> from="ANY", to="CompressedIntegerList"
#> from="ANY", to="CompressedList"
#> from="ANY", to="CompressedLogicalList"
#> from="ANY", to="CompressedNumericList"
#> from="ANY", to="CompressedRawList"
#> from="ANY", to="CompressedRleList"
#> from="ANY", to="CompressedSplitDataFrameList"
#> from="ANY", to="DataFrame"
#> from="ANY", to="DataFrameList"
#> from="ANY", to="DataTable_OR_NULL"
#> from="ANY", to="denseMatrix"
#> from="ANY", to="DNAStringSet"
#> from="ANY", to="environment"
#> from="ANY", to="expression"
#> from="ANY", to="FactorList"
#> from="ANY", to="FilterRules"
#> from="ANY", to="function"
#> from="ANY", to="GenomicRanges"
#> from="ANY", to="GPos"
#> from="ANY", to="integer"
#> from="ANY", to="IntegerList"
#> from="ANY", to="IntegerRanges"
#> from="ANY", to="IPos"
#> from="ANY", to="IPosRanges"
#> from="ANY", to="list"
#> from="ANY", to="List"
#> from="ANY", to="logical"
#> from="ANY", to="LogicalList"
#> from="ANY", to="matrix"
#> from="ANY", to="Matrix"
#> from="ANY", to="name"
#> from="ANY", to="nsparseVector"
#> from="ANY", to="NULL"
#> from="ANY", to="numeric"
#> from="ANY", to="NumericList"
#> from="ANY", to="RawList"
#> from="ANY", to="Rle"
#> from="ANY", to="RleArray"
#> from="ANY", to="RleList"
#> from="ANY", to="RleMatrix"
#> from="ANY", to="RNAStringSet"
#> from="ANY", to="S3"
#> from="ANY", to="S4"
#> from="ANY", to="SimpleCharacterList"
#> from="ANY", to="SimpleComplexList"
#> from="ANY", to="SimpleDataFrameList"
#> from="ANY", to="SimpleFactorList"
#> from="ANY", to="SimpleIntegerList"
#> from="ANY", to="SimpleList"
#> from="ANY", to="SimpleLogicalList"
#> from="ANY", to="SimpleNumericList"
#> from="ANY", to="SimpleRawList"
#> from="ANY", to="SimpleRleList"
#> from="ANY", to="SimpleSplitDataFrameList"
#> from="ANY", to="SparseArraySeed"
#> from="ANY", to="sparseMatrix"
#> from="ANY", to="sparseVector"
#> from="ANY", to="SplitDataFrameList"
#> from="ANY", to="ts"
#> from="ANY", to="vector"
#> from="ANY", to="XStringSet"
#> from="ApplyPileupsParam", to="list"
#> from="arrayRealizationSink", to="DelayedArray"
#> from="AsIs", to="DataFrame"
#> from="AssaysInEnv", to="SimpleList"
#> from="AtomicList", to="RleViews"
#> from="atomicVector", to="dsparseVector"
#> from="atomicVector", to="sparseVector"
#> from="BasicTrackLine", to="character"
#> from="BasicTrackLine", to="GraphTrackLine"
#> from="Bed15TrackLine", to="character"
#> from="BitwiseValue", to="numeric"
#> from="BString", to="IlluminaQuality"
#> from="BString", to="MaskedBString"
#> from="BString", to="PhredQuality"
#> from="BString", to="SolexaQuality"
#> from="BStringSet", to="IlluminaQuality"
#> from="BStringSet", to="PhredQuality"
#> from="BStringSet", to="SolexaQuality"
#> from="BunchKaufman", to="lMatrix"
#> from="character", to="AAMultipleAlignment"
#> from="character", to="AAString"
#> from="character", to="BasicTrackLine"
#> from="character", to="Bed15TrackLine"
#> from="character", to="BString"
#> from="character", to="curl_closepolicy"
#> from="character", to="curl_ftpauth"
#> from="character", to="curl_ftpccc"
#> from="character", to="curl_ftpcreatedir"
#> from="character", to="curl_ftpmethod"
#> from="character", to="curl_infotype"
#> from="character", to="CURL_NETRC_OPTION"
#> from="character", to="curl_proxytype"
#> from="character", to="curl_TimeCond"
#> from="character", to="curl_usessl"
#> from="character", to="CURLcode"
#> from="character", to="CURLFORMcode"
#> from="character", to="Currency"
#> from="character", to="DNAMultipleAlignment"
#> from="character", to="DNAString"
#> from="character", to="FormattedInteger"
#> from="character", to="FormattedNumber"
#> from="character", to="GRanges"
#> from="character", to="GraphTrackLine"
#> from="character", to="IDate"
#> from="character", to="IlluminaQuality"
#> from="character", to="integer64"
#> from="character", to="IRanges"
#> from="character", to="ITime"
#> from="character", to="LLint"
#> from="character", to="Percent"
#> from="character", to="PhredQuality"
#> from="character", to="Quickload"
#> from="character", to="RNAMultipleAlignment"
#> from="character", to="RNAString"
#> from="character", to="signature"
#> from="character", to="SolexaQuality"
#> from="character", to="SuperClassMethod"
#> from="character", to="TrackLine"
#> from="character", to="XMLCodeDoc"
#> from="character", to="XMLCodeFile"
#> from="character", to="XMLNamespaceDefinitions"
#> from="character", to="XString"
#> from="CHMfactor", to="Matrix"
#> from="CHMfactor", to="pMatrix"
#> from="CHMfactor", to="sparseMatrix"
#> from="CHMfactor", to="triangularMatrix"
#> from="Cholesky", to="lMatrix"
#> from="ChunkedRleArraySeed", to="SolidRleArraySeed"
#> from="color", to="HLS"
#> from="color", to="HSV"
#> from="color", to="LAB"
#> from="color", to="LUV"
#> from="color", to="polarLAB"
#> from="color", to="polarLUV"
#> from="color", to="RGB"
#> from="color", to="sRGB"
#> from="color", to="XYZ"
#> from="CompressedAtomicList", to="list"
#> from="CompressedIRangesList", to="CompressedNormalIRangesList"
#> from="CompressedRleList", to="CompressedIRangesList"
#> from="CsparseMatrix", to="denseMatrix"
#> from="CsparseMatrix", to="lMatrix"
#> from="CsparseMatrix", to="lsparseMatrix"
#> from="CsparseMatrix", to="matrix.coo"
#> from="CsparseMatrix", to="matrix.csc"
#> from="CsparseMatrix", to="matrix.csr"
#> from="CsparseMatrix", to="nMatrix"
#> from="CsparseMatrix", to="nsparseMatrix"
#> from="CsparseMatrix", to="sparseVector"
#> from="CsparseMatrix", to="symmetricMatrix"
#> from="CsparseMatrix", to="TsparseMatrix"
#> from="data.frame", to="AnnotatedDataFrame"
#> from="data.frame", to="data.table"
#> from="data.frame", to="DataFrame"
#> from="data.frame", to="dgCMatrix"
#> from="data.frame", to="dgeMatrix"
#> from="data.frame", to="GRanges"
#> from="data.frame", to="RangedData"
#> from="data.frame", to="Seqinfo"
#> from="data.frame", to="tbl_df"
#> from="data.table", to="data.frame"
#> from="data.table", to="DataFrame"
#> from="DataFrame", to="data.frame"
#> from="DataFrame", to="data.table"
#> from="DataFrame", to="GRanges"
#> from="DataFrame", to="Grouping"
#> from="DataFrame", to="RleArray"
#> from="DataFrame", to="Seqinfo"
#> from="DataFrame", to="SplitDataFrameList"
#> from="DataFrame", to="tbl_df"
#> from="DataFrameList", to="DataFrame"
#> from="DataTable", to="RangedData"
#> from="ddenseMatrix", to="dgeMatrix"
#> from="ddenseMatrix", to="matrix"
#> from="ddiMatrix", to="CsparseMatrix"
#> from="ddiMatrix", to="ddenseMatrix"
#> from="ddiMatrix", to="dgeMatrix"
#> from="ddiMatrix", to="dsparseMatrix"
#> from="ddiMatrix", to="matrix"
#> from="ddiMatrix", to="symmetricMatrix"
#> from="ddiMatrix", to="triangularMatrix"
#> from="ddiMatrix", to="TsparseMatrix"
#> from="DelayedArray", to="DelayedMatrix"
#> from="DelayedArray", to="RleArray"
#> from="DelayedArray", to="SparseArraySeed"
#> from="DelayedMatrix", to="DataFrame"
#> from="DelayedMatrix", to="DelayedArray"
#> from="DelayedMatrix", to="dgCMatrix"
#> from="DelayedMatrix", to="RleMatrix"
#> from="DelayedMatrix", to="sparseMatrix"
#> from="denseMatrix", to="CsparseMatrix"
#> from="denseMatrix", to="generalMatrix"
#> from="denseMatrix", to="RsparseMatrix"
#> from="denseMatrix", to="symmetricMatrix"
#> from="denseMatrix", to="TsparseMatrix"
#> from="derivedDefaultMethod", to="function"
#> from="DESeqAnalysis", to="DESeqDataSet"
#> from="DESeqAnalysis", to="DESeqResults"
#> from="DESeqAnalysis", to="DESeqTransform"
#> from="dgCMatrix", to="dgeMatrix"
#> from="dgCMatrix", to="dgTMatrix"
#> from="dgCMatrix", to="dsCMatrix"
#> from="dgCMatrix", to="dtCMatrix"
#> from="dgCMatrix", to="lgCMatrix"
#> from="dgCMatrix", to="matrix"
#> from="dgCMatrix", to="matrix.csc"
#> from="dgCMatrix", to="ngCMatrix"
#> from="dgCMatrix", to="SparseArraySeed"
#> from="dgCMatrix", to="vector"
#> from="dgeMatrix", to="dgCMatrix"
#> from="dgeMatrix", to="dgTMatrix"
#> from="dgeMatrix", to="dspMatrix"
#> from="dgeMatrix", to="dsTMatrix"
#> from="dgeMatrix", to="dsyMatrix"
#> from="dgeMatrix", to="dtrMatrix"
#> from="dgeMatrix", to="lgeMatrix"
#> from="dgeMatrix", to="matrix"
#> from="dgeMatrix", to="triangularMatrix"
#> from="dgRMatrix", to="matrix.csr"
#> from="dgTMatrix", to="dgCMatrix"
#> from="dgTMatrix", to="dgeMatrix"
#> from="dgTMatrix", to="dsTMatrix"
#> from="dgTMatrix", to="dtCMatrix"
#> from="dgTMatrix", to="dtTMatrix"
#> from="dgTMatrix", to="matrix"
#> from="dgTMatrix", to="matrix.coo"
#> from="dgTMatrix", to="symmetricMatrix"
#> from="dgTMatrix", to="triangularMatrix"
#> from="diagonalMatrix", to="denseMatrix"
#> from="diagonalMatrix", to="generalMatrix"
#> from="diagonalMatrix", to="matrix"
#> from="diagonalMatrix", to="nMatrix"
#> from="diagonalMatrix", to="nsparseMatrix"
#> from="diagonalMatrix", to="sparseVector"
#> from="dMatrix", to="lMatrix"
#> from="dMatrix", to="nMatrix"
#> from="DNAString", to="MaskedDNAString"
#> from="Doctype", to="character"
#> from="dpoMatrix", to="corMatrix"
#> from="dpoMatrix", to="dppMatrix"
#> from="dpoMatrix", to="lMatrix"
#> from="dpoMatrix", to="nMatrix"
#> from="dppMatrix", to="CsparseMatrix"
#> from="dppMatrix", to="dpoMatrix"
#> from="dppMatrix", to="lMatrix"
#> from="dppMatrix", to="nMatrix"
#> from="dppMatrix", to="sparseMatrix"
#> from="dsCMatrix", to="denseMatrix"
#> from="dsCMatrix", to="dgCMatrix"
#> from="dsCMatrix", to="dgeMatrix"
#> from="dsCMatrix", to="dgTMatrix"
#> from="dsCMatrix", to="dsRMatrix"
#> from="dsCMatrix", to="dsTMatrix"
#> from="dsCMatrix", to="dsyMatrix"
#> from="dsCMatrix", to="generalMatrix"
#> from="dsCMatrix", to="lsCMatrix"
#> from="dsCMatrix", to="matrix"
#> from="dsCMatrix", to="nsCMatrix"
#> from="dsCMatrix", to="vector"
#> from="dsparseMatrix", to="matrix.csr"
#> from="dspMatrix", to="CsparseMatrix"
#> from="dspMatrix", to="dpoMatrix"
#> from="dspMatrix", to="dppMatrix"
#> from="dspMatrix", to="dsyMatrix"
#> from="dspMatrix", to="lspMatrix"
#> from="dspMatrix", to="matrix"
#> from="dspMatrix", to="sparseMatrix"
#> from="dsTMatrix", to="dgeMatrix"
#> from="dsTMatrix", to="dgTMatrix"
#> from="dsTMatrix", to="dsCMatrix"
#> from="dsTMatrix", to="dsyMatrix"
#> from="dsTMatrix", to="lsTMatrix"
#> from="dsTMatrix", to="matrix"
#> from="dsyMatrix", to="corMatrix"
#> from="dsyMatrix", to="dpoMatrix"
#> from="dsyMatrix", to="dsCMatrix"
#> from="dsyMatrix", to="dspMatrix"
#> from="dsyMatrix", to="dsTMatrix"
#> from="dsyMatrix", to="lsyMatrix"
#> from="dsyMatrix", to="matrix"
#> from="dtCMatrix", to="denseMatrix"
#> from="dtCMatrix", to="dgCMatrix"
#> from="dtCMatrix", to="dgeMatrix"
#> from="dtCMatrix", to="dgTMatrix"
#> from="dtCMatrix", to="dsCMatrix"
#> from="dtCMatrix", to="dtrMatrix"
#> from="dtCMatrix", to="dtTMatrix"
#> from="dtCMatrix", to="ltCMatrix"
#> from="dtCMatrix", to="matrix"
#> from="dtCMatrix", to="ntCMatrix"
#> from="dtpMatrix", to="dtrMatrix"
#> from="dtpMatrix", to="dtTMatrix"
#> from="dtpMatrix", to="ltpMatrix"
#> from="dtpMatrix", to="matrix"
#> from="dtrMatrix", to="CsparseMatrix"
#> from="dtrMatrix", to="dtpMatrix"
#> from="dtrMatrix", to="ltrMatrix"
#> from="dtrMatrix", to="matrix"
#> from="dtrMatrix", to="sparseMatrix"
#> from="dtTMatrix", to="dgeMatrix"
#> from="dtTMatrix", to="dgTMatrix"
#> from="dtTMatrix", to="dtCMatrix"
#> from="dtTMatrix", to="dtrMatrix"
#> from="dtTMatrix", to="generalMatrix"
#> from="dtTMatrix", to="matrix"
#> from="eSet", to="ExpressionSet"
#> from="eSet", to="MultiSet"
#> from="ExpressionSet", to="data.frame"
#> from="ExpressionSet", to="RangedSummarizedExperiment"
#> from="ExpressionSet", to="SummarizedExperiment"
#> from="exprSet", to="ExpressionSet"
#> from="factor", to="GRanges"
#> from="factor", to="IRanges"
#> from="factor", to="ManyToOneGrouping"
#> from="factor", to="sparseMatrix"
#> from="FactorList", to="Grouping"
#> from="function", to="classGeneratorFunction"
#> from="function", to="FilterClosure"
#> from="function", to="OptionalFunction"
#> from="GAlignmentPairs", to="DataFrame"
#> from="GAlignmentPairs", to="GAlignments"
#> from="GAlignmentPairs", to="GAlignmentsList"
#> from="GAlignmentPairs", to="GRanges"
#> from="GAlignmentPairs", to="GRangesList"
#> from="GAlignmentPairs", to="IntegerRanges"
#> from="GAlignments", to="DataFrame"
#> from="GAlignments", to="GRanges"
#> from="GAlignments", to="GRangesList"
#> from="GAlignments", to="IntegerRanges"
#> from="GAlignments", to="IntegerRangesList"
#> from="GAlignmentsList", to="GAlignmentPairs"
#> from="GAlignmentsList", to="GRanges"
#> from="GAlignmentsList", to="GRangesList"
#> from="GAlignmentsList", to="IntegerRanges"
#> from="GAlignmentsList", to="IntegerRangesList"
#> from="GenomicRanges", to="BigWigSelection"
#> from="GenomicRanges", to="CompressedGRangesList"
#> from="GenomicRanges", to="CompressedIRangesList"
#> from="GenomicRanges", to="GAlignments"
#> from="GenomicRanges", to="GNCList"
#> from="GenomicRanges", to="GRanges"
#> from="GenomicRanges", to="GRangesList"
#> from="GenomicRanges", to="Grouping"
#> from="GenomicRanges", to="IntegerRangesList"
#> from="GenomicRanges", to="IRangesList"
#> from="GenomicRanges", to="RangedData"
#> from="GNCList", to="GRanges"
#> from="GPos", to="GRanges"
#> from="GRanges", to="data.table"
#> from="GRanges", to="GPos"
#> from="GRanges", to="tbl_df"
#> from="GRanges", to="UCSCData"
#> from="GRangesList", to="CompressedIRangesList"
#> from="GRangesList", to="IntegerRangesList"
#> from="GRangesList", to="IRangesList"
#> from="graph", to="CsparseMatrix"
#> from="graph", to="Matrix"
#> from="graphAM", to="sparseMatrix"
#> from="graphNEL", to="CsparseMatrix"
#> from="graphNEL", to="TsparseMatrix"
#> from="GraphTrackLine", to="BasicTrackLine"
#> from="GraphTrackLine", to="character"
#> from="groupGenericFunction", to="genericFunction"
#> from="grouping", to="Grouping"
#> from="grouping", to="ManyToOneGrouping"
#> from="Hits", to="CompressedIntegerList"
#> from="Hits", to="DataFrame"
#> from="Hits", to="Grouping"
#> from="Hits", to="IntegerList"
#> from="Hits", to="List"
#> from="Hits", to="SortedByQueryHits"
#> from="indMatrix", to="CsparseMatrix"
#> from="indMatrix", to="dMatrix"
#> from="indMatrix", to="dsparseMatrix"
#> from="indMatrix", to="lMatrix"
#> from="indMatrix", to="lsparseMatrix"
#> from="indMatrix", to="matrix"
#> from="indMatrix", to="ngeMatrix"
#> from="indMatrix", to="ngTMatrix"
#> from="indMatrix", to="nMatrix"
#> from="indMatrix", to="nsparseMatrix"
#> from="indMatrix", to="TsparseMatrix"
#> from="integer", to="curl_closepolicy"
#> from="integer", to="curl_ftpauth"
#> from="integer", to="curl_ftpccc"
#> from="integer", to="curl_ftpcreatedir"
#> from="integer", to="curl_ftpmethod"
#> from="integer", to="curl_infotype"
#> from="integer", to="CURL_NETRC_OPTION"
#> from="integer", to="curl_proxytype"
#> from="integer", to="curl_TimeCond"
#> from="integer", to="curl_usessl"
#> from="integer", to="CURLcode"
#> from="integer", to="CURLFORMcode"
#> from="integer", to="IlluminaQuality"
#> from="integer", to="indMatrix"
#> from="integer", to="IRanges"
#> from="integer", to="List"
#> from="integer", to="LLint"
#> from="integer", to="NormalIRanges"
#> from="integer", to="PhredQuality"
#> from="integer", to="pMatrix"
#> from="integer", to="SolexaQuality"
#> from="integer", to="XVector"
#> from="integer64", to="character"
#> from="IntegerList", to="IlluminaQuality"
#> from="IntegerList", to="PhredQuality"
#> from="IntegerList", to="SolexaQuality"
#> from="IntegerRanges", to="CompressedIPosList"
#> from="IntegerRanges", to="CompressedIRangesList"
#> from="IntegerRanges", to="IPos"
#> from="IntegerRanges", to="IPosList"
#> from="IntegerRanges", to="IRanges"
#> from="IntegerRanges", to="NCList"
#> from="IntegerRanges", to="PartitioningByEnd"
#> from="IntegerRanges", to="PartitioningByWidth"
#> from="IntegerRanges", to="RangedData"
#> from="IntegerRangesList", to="BigWigSelection"
#> from="IntegerRangesList", to="CompressedNormalIRangesList"
#> from="IntegerRangesList", to="GRanges"
#> from="IntegerRangesList", to="NCLists"
#> from="IntegerRangesList", to="NormalIRangesList"
#> from="IntegerRangesList", to="RangedData"
#> from="IntegerRangesList", to="RangedSelection"
#> from="IntegerRangesList", to="SimpleIRangesList"
#> from="IntegerRangesList", to="SimpleNormalIRangesList"
#> from="IPosRanges", to="CompressedIntegerList"
#> from="IPosRanges", to="IntegerList"
#> from="IRanges", to="IPosList"
#> from="IRanges", to="NormalIRanges"
#> from="ldenseMatrix", to="matrix"
#> from="ldiMatrix", to="CsparseMatrix"
#> from="ldiMatrix", to="ldenseMatrix"
#> from="ldiMatrix", to="lsparseMatrix"
#> from="ldiMatrix", to="symmetricMatrix"
#> from="ldiMatrix", to="triangularMatrix"
#> from="ldiMatrix", to="TsparseMatrix"
#> from="lgCMatrix", to="dgCMatrix"
#> from="lgCMatrix", to="lgeMatrix"
#> from="lgCMatrix", to="lgTMatrix"
#> from="lgCMatrix", to="ltCMatrix"
#> from="lgCMatrix", to="matrix"
#> from="lgeMatrix", to="dgeMatrix"
#> from="lgeMatrix", to="lgCMatrix"
#> from="lgeMatrix", to="lgTMatrix"
#> from="lgeMatrix", to="lspMatrix"
#> from="lgeMatrix", to="lsyMatrix"
#> from="lgeMatrix", to="ltpMatrix"
#> from="lgeMatrix", to="ltrMatrix"
#> from="lgeMatrix", to="matrix"
#> from="lgeMatrix", to="triangularMatrix"
#> from="lgTMatrix", to="dgTMatrix"
#> from="lgTMatrix", to="lgCMatrix"
#> from="lgTMatrix", to="lgeMatrix"
#> from="lgTMatrix", to="lsCMatrix"
#> from="lgTMatrix", to="ltTMatrix"
#> from="lgTMatrix", to="matrix"
#> from="lgTMatrix", to="symmetricMatrix"
#> from="lgTMatrix", to="triangularMatrix"
#> from="libxmlTypeTable", to="list"
#> from="list", to="CharacterList"
#> from="list", to="ComplexList"
#> from="list", to="CompressedGRangesList"
#> from="list", to="CompressedIRangesList"
#> from="List", to="CompressedIRangesList"
#> from="List", to="data.frame"
#> from="list", to="DataFrame"
#> from="list", to="DataFrameList"
#> from="list", to="FactorList"
#> from="list", to="FGSEAList"
#> from="list", to="GAlignmentsList"
#> from="list", to="GRangesList"
#> from="list", to="indMatrix"
#> from="list", to="IntegerList"
#> from="list", to="IRangesList"
#> from="List", to="IRangesList"
#> from="list", to="List"
#> from="List", to="list"
#> from="list", to="LogicalList"
#> from="list", to="NumericList"
#> from="list", to="RangedData"
#> from="list", to="RawList"
#> from="list", to="RleList"
#> from="list", to="SimpleIRangesList"
#> from="List", to="SimpleIRangesList"
#> from="list", to="SplitDataFrameList"
#> from="lMatrix", to="dgCMatrix"
#> from="lMatrix", to="dMatrix"
#> from="lMatrix", to="nMatrix"
#> from="logical", to="abIndex"
#> from="logical", to="IRanges"
#> from="logical", to="LLint"
#> from="logical", to="NormalIRanges"
#> from="LogicalList", to="CompressedNormalIRangesList"
#> from="LogicalList", to="NormalIRangesList"
#> from="LogicalList", to="SimpleNormalIRangesList"
#> from="lsCMatrix", to="dsCMatrix"
#> from="lsCMatrix", to="generalMatrix"
#> from="lsCMatrix", to="lgCMatrix"
#> from="lsCMatrix", to="lgTMatrix"
#> from="lsCMatrix", to="lsTMatrix"
#> from="lsCMatrix", to="matrix"
#> from="lsparseMatrix", to="dsparseMatrix"
#> from="lsparseMatrix", to="matrix"
#> from="lspMatrix", to="dspMatrix"
#> from="lspMatrix", to="lgeMatrix"
#> from="lspMatrix", to="lsyMatrix"
#> from="lsTMatrix", to="lgCMatrix"
#> from="lsTMatrix", to="lgTMatrix"
#> from="lsTMatrix", to="lsCMatrix"
#> from="lsTMatrix", to="lsyMatrix"
#> from="lsTMatrix", to="matrix"
#> from="lsyMatrix", to="dsyMatrix"
#> from="lsyMatrix", to="lgeMatrix"
#> from="lsyMatrix", to="lspMatrix"
#> from="ltCMatrix", to="dMatrix"
#> from="ltCMatrix", to="lgCMatrix"
#> from="ltCMatrix", to="ltTMatrix"
#> from="ltCMatrix", to="matrix"
#> from="ltpMatrix", to="dtpMatrix"
#> from="ltpMatrix", to="lgeMatrix"
#> from="ltpMatrix", to="ltrMatrix"
#> from="ltrMatrix", to="dtrMatrix"
#> from="ltrMatrix", to="lgeMatrix"
#> from="ltrMatrix", to="ltpMatrix"
#> from="ltTMatrix", to="dtTMatrix"
#> from="ltTMatrix", to="generalMatrix"
#> from="ltTMatrix", to="lgCMatrix"
#> from="ltTMatrix", to="lgTMatrix"
#> from="ltTMatrix", to="ltCMatrix"
#> from="ltTMatrix", to="ltrMatrix"
#> from="ltTMatrix", to="matrix"
#> from="ManyToOneGrouping", to="factor"
#> from="MaskCollection", to="NormalIRanges"
#> from="MaskedAAString", to="AAString"
#> from="MaskedBString", to="BString"
#> from="MaskedDNAString", to="DNAString"
#> from="MaskedRNAString", to="RNAString"
#> from="MaskedXString", to="MaskCollection"
#> from="MaskedXString", to="MaskedAAString"
#> from="MaskedXString", to="MaskedBString"
#> from="MaskedXString", to="MaskedDNAString"
#> from="MaskedXString", to="MaskedRNAString"
#> from="MaskedXString", to="NormalIRanges"
#> from="MaskedXString", to="Views"
#> from="MaskedXString", to="XStringViews"
#> from="matrix.coo", to="CsparseMatrix"
#> from="matrix.coo", to="dgCMatrix"
#> from="matrix.coo", to="dgTMatrix"
#> from="matrix.coo", to="Matrix"
#> from="matrix.coo", to="TsparseMatrix"
#> from="matrix.csc", to="CsparseMatrix"
#> from="matrix.csc", to="dgCMatrix"
#> from="matrix.csc", to="Matrix"
#> from="matrix.csc", to="TsparseMatrix"
#> from="matrix.csr", to="CsparseMatrix"
#> from="matrix.csr", to="dgCMatrix"
#> from="matrix.csr", to="dgRMatrix"
#> from="matrix.csr", to="Matrix"
#> from="matrix.csr", to="RsparseMatrix"
#> from="matrix.csr", to="TsparseMatrix"
#> from="Matrix", to="complex"
#> from="matrix", to="corMatrix"
#> from="Matrix", to="corMatrix"
#> from="matrix", to="CsparseMatrix"
#> from="Matrix", to="CsparseMatrix"
#> from="matrix", to="data.frame"
#> from="Matrix", to="data.frame"
#> from="Matrix", to="denseMatrix"
#> from="matrix", to="dgCMatrix"
#> from="matrix", to="dgeMatrix"
#> from="matrix", to="dgRMatrix"
#> from="matrix", to="dgTMatrix"
#> from="matrix", to="diagonalMatrix"
#> from="Matrix", to="diagonalMatrix"
#> from="matrix", to="dpoMatrix"
#> from="Matrix", to="dpoMatrix"
#> from="matrix", to="dppMatrix"
#> from="Matrix", to="dppMatrix"
#> from="matrix", to="dsCMatrix"
#> from="matrix", to="dspMatrix"
#> from="matrix", to="dsTMatrix"
#> from="matrix", to="dsyMatrix"
#> from="matrix", to="dtCMatrix"
#> from="matrix", to="dtpMatrix"
#> from="matrix", to="dtrMatrix"
#> from="matrix", to="dtTMatrix"
#> from="matrix", to="indMatrix"
#> from="Matrix", to="integer"
#> from="matrix", to="ldenseMatrix"
#> from="matrix", to="lgCMatrix"
#> from="matrix", to="lgeMatrix"
#> from="matrix", to="lgTMatrix"
#> from="matrix", to="lMatrix"
#> from="Matrix", to="logical"
#> from="matrix", to="lsCMatrix"
#> from="matrix", to="lspMatrix"
#> from="matrix", to="lsyMatrix"
#> from="matrix", to="ltCMatrix"
#> from="matrix", to="ltpMatrix"
#> from="matrix", to="ltrMatrix"
#> from="matrix", to="ltTMatrix"
#> from="matrix", to="Matrix"
#> from="Matrix", to="matrix"
#> from="matrix", to="ndenseMatrix"
#> from="matrix", to="ngCMatrix"
#> from="matrix", to="ngeMatrix"
#> from="matrix", to="ngTMatrix"
#> from="matrix", to="nMatrix"
#> from="matrix", to="nspMatrix"
#> from="matrix", to="nsyMatrix"
#> from="matrix", to="ntCMatrix"
#> from="matrix", to="ntpMatrix"
#> from="matrix", to="ntrMatrix"
#> from="matrix", to="ntTMatrix"
#> from="Matrix", to="numeric"
#> from="matrix", to="pMatrix"
#> from="matrix", to="RsparseMatrix"
#> from="Matrix", to="sparseMatrix"
#> from="matrix", to="symmetricMatrix"
#> from="matrix", to="triangularMatrix"
#> from="matrix", to="TsparseMatrix"
#> from="Matrix", to="vector"
#> from="MIndex", to="CompressedIRangesList"
#> from="MulticoreParam", to="BiocParallelParam"
#> from="MulticoreParam", to="SnowParam"
#> from="MultipleAlignment", to="AAStringSet"
#> from="MultipleAlignment", to="BStringSet"
#> from="MultipleAlignment", to="DNAStringSet"
#> from="MultipleAlignment", to="RNAStringSet"
#> from="NCLists", to="CompressedIRangesList"
#> from="NCLists", to="IRangesList"
#> from="ndenseMatrix", to="CsparseMatrix"
#> from="ndenseMatrix", to="ldenseMatrix"
#> from="ndenseMatrix", to="matrix"
#> from="ndenseMatrix", to="nsparseMatrix"
#> from="ndenseMatrix", to="sparseMatrix"
#> from="ndenseMatrix", to="TsparseMatrix"
#> from="ngCMatrix", to="dgCMatrix"
#> from="ngCMatrix", to="dMatrix"
#> from="ngCMatrix", to="dsparseMatrix"
#> from="ngCMatrix", to="lgCMatrix"
#> from="ngCMatrix", to="lMatrix"
#> from="ngCMatrix", to="lsparseMatrix"
#> from="ngCMatrix", to="matrix"
#> from="ngCMatrix", to="ntCMatrix"
#> from="ngeMatrix", to="dgeMatrix"
#> from="ngeMatrix", to="lgeMatrix"
#> from="ngeMatrix", to="matrix"
#> from="ngeMatrix", to="ngCMatrix"
#> from="ngeMatrix", to="ngTMatrix"
#> from="ngeMatrix", to="nspMatrix"
#> from="ngeMatrix", to="nsyMatrix"
#> from="ngeMatrix", to="ntpMatrix"
#> from="ngeMatrix", to="ntrMatrix"
#> from="ngeMatrix", to="triangularMatrix"
#> from="ngTMatrix", to="dgTMatrix"
#> from="ngTMatrix", to="dMatrix"
#> from="ngTMatrix", to="dsparseMatrix"
#> from="ngTMatrix", to="generalMatrix"
#> from="ngTMatrix", to="lgeMatrix"
#> from="ngTMatrix", to="lgTMatrix"
#> from="ngTMatrix", to="lMatrix"
#> from="ngTMatrix", to="matrix"
#> from="ngTMatrix", to="ngCMatrix"
#> from="ngTMatrix", to="ngeMatrix"
#> from="ngTMatrix", to="ntTMatrix"
#> from="ngTMatrix", to="symmetricMatrix"
#> from="ngTMatrix", to="triangularMatrix"
#> from="nMatrix", to="dMatrix"
#> from="nMatrix", to="indMatrix"
#> from="nMatrix", to="lMatrix"
#> from="nMatrix", to="pMatrix"
#> from="NormalIRangesList", to="CompressedNormalIRangesList"
#> from="nsCMatrix", to="dMatrix"
#> from="nsCMatrix", to="dsCMatrix"
#> from="nsCMatrix", to="dsparseMatrix"
#> from="nsCMatrix", to="generalMatrix"
#> from="nsCMatrix", to="lMatrix"
#> from="nsCMatrix", to="lsCMatrix"
#> from="nsCMatrix", to="lsparseMatrix"
#> from="nsCMatrix", to="matrix"
#> from="nsCMatrix", to="ngCMatrix"
#> from="nsCMatrix", to="nsTMatrix"
#> from="nsparseMatrix", to="dsparseMatrix"
#> from="nsparseVector", to="dsparseVector"
#> from="nsparseVector", to="isparseVector"
#> from="nsparseVector", to="lsparseVector"
#> from="nsparseVector", to="zsparseVector"
#> from="nspMatrix", to="dspMatrix"
#> from="nspMatrix", to="lspMatrix"
#> from="nspMatrix", to="ngeMatrix"
#> from="nspMatrix", to="nsyMatrix"
#> from="nsTMatrix", to="dsTMatrix"
#> from="nsTMatrix", to="matrix"
#> from="nsTMatrix", to="ngCMatrix"
#> from="nsTMatrix", to="ngTMatrix"
#> from="nsTMatrix", to="nsCMatrix"
#> from="nsTMatrix", to="nsyMatrix"
#> from="nsyMatrix", to="dsyMatrix"
#> from="nsyMatrix", to="lsyMatrix"
#> from="nsyMatrix", to="ngeMatrix"
#> from="nsyMatrix", to="nspMatrix"
#> from="ntCMatrix", to="dMatrix"
#> from="ntCMatrix", to="dsparseMatrix"
#> from="ntCMatrix", to="dtCMatrix"
#> from="ntCMatrix", to="lMatrix"
#> from="ntCMatrix", to="lsparseMatrix"
#> from="ntCMatrix", to="ltCMatrix"
#> from="ntCMatrix", to="matrix"
#> from="ntCMatrix", to="ngCMatrix"
#> from="ntCMatrix", to="TsparseMatrix"
#> from="ntpMatrix", to="dtpMatrix"
#> from="ntpMatrix", to="ltpMatrix"
#> from="ntpMatrix", to="ngeMatrix"
#> from="ntpMatrix", to="ntrMatrix"
#> from="ntrMatrix", to="dtrMatrix"
#> from="ntrMatrix", to="ltrMatrix"
#> from="ntrMatrix", to="ngeMatrix"
#> from="ntrMatrix", to="ntpMatrix"
#> from="ntTMatrix", to="dtTMatrix"
#> from="ntTMatrix", to="generalMatrix"
#> from="ntTMatrix", to="matrix"
#> from="ntTMatrix", to="ngCMatrix"
#> from="ntTMatrix", to="ngTMatrix"
#> from="ntTMatrix", to="ntCMatrix"
#> from="ntTMatrix", to="ntrMatrix"
#> from="NULL", to="DataFrame"
#> from="NULL", to="OptionalFunction"
#> from="NULL", to="optionalMethod"
#> from="NULL", to="XMLNamespaceDefinitions"
#> from="NULLcluster", to="cluster"
#> from="numeric", to="abIndex"
#> from="numeric", to="CsparseMatrix"
#> from="numeric", to="curl_closepolicy"
#> from="numeric", to="curl_ftpauth"
#> from="numeric", to="curl_ftpccc"
#> from="numeric", to="curl_ftpcreatedir"
#> from="numeric", to="curl_ftpmethod"
#> from="numeric", to="curl_infotype"
#> from="numeric", to="CURL_NETRC_OPTION"
#> from="numeric", to="curl_proxytype"
#> from="numeric", to="curl_TimeCond"
#> from="numeric", to="curl_usessl"
#> from="numeric", to="CURLcode"
#> from="numeric", to="CURLFORMcode"
#> from="numeric", to="IlluminaQuality"
#> from="numeric", to="indMatrix"
#> from="numeric", to="IRanges"
#> from="numeric", to="LLint"
#> from="numeric", to="NetrcEnum"
#> from="numeric", to="NormalIRanges"
#> from="numeric", to="PhredQuality"
#> from="numeric", to="pMatrix"
#> from="numeric", to="seqMat"
#> from="numeric", to="SolexaQuality"
#> from="numeric", to="TsparseMatrix"
#> from="numeric", to="XDouble"
#> from="numeric", to="XInteger"
#> from="numeric", to="XRaw"
#> from="numeric", to="XVector"
#> from="NumericList", to="IlluminaQuality"
#> from="NumericList", to="PhredQuality"
#> from="NumericList", to="SolexaQuality"
#> from="numLike", to="dgeMatrix"
#> from="oldClass", to="S3"
#> from="Pairs", to="DataFrame"
#> from="PartitioningByEnd", to="PartitioningMap"
#> from="pBunchKaufman", to="lMatrix"
#> from="pCholesky", to="lMatrix"
#> from="phenoData", to="AnnotatedDataFrame"
#> from="RangedData", to="CompressedIRangesList"
#> from="RangedData", to="DataFrame"
#> from="RangedData", to="GRanges"
#> from="RangedData", to="IRangesList"
#> from="RangedSummarizedExperiment", to="ExpressionSet"
#> from="RangedSummarizedExperiment", to="SingleCellExperiment"
#> from="RangedSummarizedExperiment", to="SummarizedExperiment"
#> from="raw", to="XRaw"
#> from="raw", to="XVector"
#> from="RCurlBinaryBuffer", to="raw"
#> from="Rle", to="character"
#> from="Rle", to="complex"
#> from="Rle", to="data.frame"
#> from="Rle", to="factor"
#> from="Rle", to="integer"
#> from="Rle", to="IRanges"
#> from="Rle", to="list"
#> from="Rle", to="logical"
#> from="Rle", to="NormalIRanges"
#> from="Rle", to="numeric"
#> from="Rle", to="RangedData"
#> from="Rle", to="raw"
#> from="Rle", to="vector"
#> from="RleArray", to="Rle"
#> from="RleArray", to="RleMatrix"
#> from="RleList", to="CompressedNormalIRangesList"
#> from="RleList", to="GRanges"
#> from="RleList", to="NormalIRangesList"
#> from="RleList", to="RangedData"
#> from="RleList", to="RleArray"
#> from="RleList", to="SimpleNormalIRangesList"
#> from="RleMatrix", to="DataFrame"
#> from="RleMatrix", to="RleList"
#> from="RleRealizationSink", to="ChunkedRleArraySeed"
#> from="RleRealizationSink", to="DelayedArray"
#> from="RleRealizationSink", to="Rle"
#> from="RleRealizationSink", to="RleArray"
#> from="RleViewsList", to="GRanges"
#> from="RleViewsList", to="IRangesList"
#> from="RleViewsList", to="RangedData"
#> from="RleViewsList", to="SimpleIRangesList"
#> from="RNAString", to="MaskedRNAString"
#> from="RsparseMatrix", to="CsparseMatrix"
#> from="RsparseMatrix", to="denseMatrix"
#> from="RsparseMatrix", to="dgeMatrix"
#> from="RsparseMatrix", to="dMatrix"
#> from="RsparseMatrix", to="dsparseMatrix"
#> from="RsparseMatrix", to="generalMatrix"
#> from="RsparseMatrix", to="lMatrix"
#> from="RsparseMatrix", to="lsparseMatrix"
#> from="RsparseMatrix", to="matrix"
#> from="RsparseMatrix", to="nMatrix"
#> from="RsparseMatrix", to="nsparseMatrix"
#> from="RsparseMatrix", to="TsparseMatrix"
#> from="Seqinfo", to="GRanges"
#> from="Seqinfo", to="IntegerRangesList"
#> from="seqMat", to="abIndex"
#> from="seqMat", to="numeric"
#> from="ShallowSimpleListAssays", to="SimpleList"
#> from="SharedVector", to="SharedVector_Pool"
#> from="SimpleIntegerRangesList", to="SimpleIRangesList"
#> from="SimpleIRangesList", to="SimpleNormalIRangesList"
#> from="SimpleList", to="AssaysInEnv"
#> from="SimpleList", to="CharacterList"
#> from="SimpleList", to="ComplexList"
#> from="SimpleList", to="DataFrame"
#> from="SimpleList", to="DataFrameList"
#> from="SimpleList", to="FactorList"
#> from="SimpleList", to="FilterRules"
#> from="SimpleList", to="IntegerList"
#> from="SimpleList", to="LogicalList"
#> from="SimpleList", to="NumericList"
#> from="SimpleList", to="RawList"
#> from="SimpleList", to="RleList"
#> from="SimpleList", to="ShallowSimpleListAssays"
#> from="SimpleList", to="SimpleIRangesList"
#> from="SimpleList", to="SplitDataFrameList"
#> from="SOCKcluster", to="SnowParam"
#> from="SolidRleArraySeed", to="Rle"
#> from="SortedByQueryHits", to="CompressedIntegerList"
#> from="SortedByQueryHits", to="IntegerList"
#> from="SortedByQueryHits", to="IntegerRanges"
#> from="SortedByQueryHits", to="IRanges"
#> from="SortedByQueryHits", to="List"
#> from="SortedByQueryHits", to="Partitioning"
#> from="SortedByQueryHits", to="PartitioningByEnd"
#> from="SparseArraySeed", to="dgCMatrix"
#> from="SparseArraySeed", to="sparseMatrix"
#> from="sparseMatrix", to="data.frame"
#> from="sparseMatrix", to="DataFrame"
#> from="sparseMatrix", to="generalMatrix"
#> from="sparseMatrix", to="graph"
#> from="sparseMatrix", to="graphNEL"
#> from="sparseMatrix", to="indMatrix"
#> from="sparseMatrix", to="RsparseMatrix"
#> from="sparseMatrix", to="sparseVector"
#> from="sparseMatrix", to="symmetricMatrix"
#> from="sparseMatrix", to="triangularMatrix"
#> from="sparseVector", to="CsparseMatrix"
#> from="sparseVector", to="integer"
#> from="sparseVector", to="logical"
#> from="sparseVector", to="Matrix"
#> from="sparseVector", to="numeric"
#> from="sparseVector", to="sparseMatrix"
#> from="sparseVector", to="TsparseMatrix"
#> from="sparseVector", to="vector"
#> from="spawnedMPIcluster", to="SnowParam"
#> from="SplitDataFrameList", to="DataFrame"
#> from="standardGeneric", to="FilterClosure"
#> from="SummarizedExperiment", to="ExpressionSet"
#> from="SummarizedExperiment", to="RangedSummarizedExperiment"
#> from="SummarizedExperiment", to="SingleCellExperiment"
#> from="table", to="DataFrame"
#> from="table", to="sparseMatrix"
#> from="tbl_df", to="DataFrame"
#> from="TrackLine", to="character"
#> from="triangularMatrix", to="symmetricMatrix"
#> from="TsparseMatrix", to="CsparseMatrix"
#> from="TsparseMatrix", to="graphNEL"
#> from="TsparseMatrix", to="lMatrix"
#> from="TsparseMatrix", to="lsparseMatrix"
#> from="TsparseMatrix", to="matrix"
#> from="TsparseMatrix", to="nMatrix"
#> from="TsparseMatrix", to="nsparseMatrix"
#> from="TsparseMatrix", to="sparseVector"
#> from="UCSCData", to="GRanges"
#> from="URI", to="character"
#> from="Vector", to="character"
#> from="Vector", to="complex"
#> from="vector", to="data.frame"
#> from="Vector", to="data.frame"
#> from="Vector", to="DataFrame"
#> from="Vector", to="factor"
#> from="vector", to="Grouping"
#> from="Vector", to="integer"
#> from="Vector", to="logical"
#> from="vector", to="ManyToManyGrouping"
#> from="vector", to="ManyToOneGrouping"
#> from="Vector", to="numeric"
#> from="Vector", to="raw"
#> from="Vector", to="vector"
#> from="Vector", to="Views"
#> from="vector", to="XMLInternalNode"
#> from="Versions", to="character"
#> from="Views", to="IntegerRanges"
#> from="Views", to="IRanges"
#> from="Views", to="NormalIRanges"
#> from="XMLAbstractDocument", to="XMLAbstractNode"
#> from="XMLAbstractNode", to="character"
#> from="XMLAbstractNode", to="Date"
#> from="XMLAbstractNode", to="integer"
#> from="XMLAbstractNode", to="logical"
#> from="XMLAbstractNode", to="numeric"
#> from="XMLAbstractNode", to="POSIXct"
#> from="XMLAbstractNode", to="URL"
#> from="XMLCodeFile", to="XMLCodeDoc"
#> from="XMLDocument", to="XMLInternalDocument"
#> from="XMLHashTreeNode", to="XMLHashTree"
#> from="XMLInternalDocument", to="character"
#> from="XMLInternalDocument", to="XMLHashTree"
#> from="XMLInternalDocument", to="XMLInternalNode"
#> from="XMLInternalDOM", to="character"
#> from="XMLInternalNode", to="character"
#> from="XMLInternalNode", to="XMLHashTree"
#> from="XMLInternalNode", to="XMLInternalDocument"
#> from="XMLInternalNode", to="XMLNode"
#> from="XMLInternalTextNode", to="character"
#> from="XMLNamespace", to="character"
#> from="XMLNamespaceDefinition", to="character"
#> from="XMLNamespaceDefinitions", to="character"
#> from="XMLNamespaceRef", to="character"
#> from="XMLNode", to="XMLInternalNode"
#> from="xsparseVector", to="dsparseVector"
#> from="xsparseVector", to="isparseVector"
#> from="xsparseVector", to="lsparseVector"
#> from="xsparseVector", to="nsparseVector"
#> from="xsparseVector", to="zsparseVector"
#> from="XString", to="AAString"
#> from="XString", to="BString"
#> from="XString", to="DNAString"
#> from="XString", to="RNAString"
#> from="XStringQuality", to="IntegerList"
#> from="XStringQuality", to="matrix"
#> from="XStringQuality", to="NumericList"
#> from="XStringSet", to="Views"
#> from="XStringSet", to="XStringViews"
#> from="XStringViews", to="AAStringSet"
#> from="XStringViews", to="BStringSet"
#> from="XStringViews", to="DNAStringSet"
#> from="XStringViews", to="RNAStringSet"
#> from="XStringViews", to="XStringSet"
#> from="xtabs", to="DataFrame"
#> from="XVector", to="Rle"
#>